backend/requirements.txt

# Produktions-Abhängigkeiten
fastapi
uvicorn
langchain
faiss-cpu
sentence-transformers
python-dotenv
# Produktions-Abhängigkeiten
fastapi
uvicorn[standard]
pydantic
python-dotenv

# LangChain & Vektordatenbank
langchain
langchain-community
langchain-core
langchain-ollama
sentence-transformers
faiss-cpu

# Hilfsbibliotheken
numpy